backend-prototipo/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
dataclasses-json==0.6.1
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.306
langsmith==0.0.41
lit==17.0.1
llvmlite==0.41.0
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.58.0
numexpr==2.8.7
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230918
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
